@@ -24,6 +24,11 @@
 			<artifactId>jaxb-core</artifactId>
 			<version>2.2.11</version>
 		</dependency>
+		<dependency>
+			<groupId>javax.xml.bind</groupId>
+			<artifactId>jaxb-api</artifactId>
+			<version>2.3.0</version>
+		</dependency>
 		<!-- TEST FRAMEWORKS -->
 		<dependency>
 			<groupId>junit</groupId>
@@ -48,10 +53,15 @@
 			<version>4.0.1-alpha</version>
 			<scope>test</scope>
 		</dependency>
+		<dependency>
+			<groupId>de.urkallinger.config</groupId>
+			<artifactId>de.urkallinger.config</artifactId>
+			<version>1.0.0</version>
+		</dependency>
 	</dependencies>
 	<properties>
-		<maven.compiler.source>1.8</maven.compiler.source>
-		<maven.compiler.target>1.8</maven.compiler.target>
+		<maven.compiler.source>9</maven.compiler.source>
+		<maven.compiler.target>9</maven.compiler.target>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<mainClass>de.urkallinger.copymanager.MainApp</mainClass>
 	</properties>
