@@ -1,6 +1,6 @@
 package de.urkallinger.copymanager.exceptions;
 
-public class CMException extends Exception {
+public class CMException extends RuntimeException {
 
 	private static final long serialVersionUID = 1L;
 
